testare unitara. code coverage

+++ Laborator12- Testare unitara/src/ro/ase/csie/cts/g1087/testare/modele/OperatiiMatematice.java
@@ -15,7 +15,7 @@ public class OperatiiMatematice {
 		{
 			if(b==0)
 				throw new ExceptieImpartireLaZero();
-			return a/b;
+			return (double)a/b;
 		}
 				
 }
